implement "leightweight" export with external dependencies
Closes datavis-tech/vizhub-issue-tracker/issues/672
