+++ packages/presenters/package.json
@@ -20,7 +20,7 @@
     "vizhub-entities": "1",
     "vizhub-libraries": "^1.8.0",
     "vizhub-use-cases": "1",
-    "xmldom": "^0.4.0"
+    "jsdom": "^16.4.0"
   },
   "devDependencies": {
     "esm": "^3.2.25",
