+++ packages/presenters/src/index.js
@@ -2,5 +2,7 @@ export { VisualizationViewModel } from './visualizationViewModel';
 export { DatasetViewModel } from './datasetViewModel';
 export { bundle } from './bundle';
 export { computeSrcDoc } from './computeSrcDoc';
+export { getLibraries } from './getLibraries';
+export { getComputedIndexHtml } from './getComputedIndexHtml';
 export * from './accessors';
 export * from './packageJson';
