+++ packages/useCases/src/interactors/exportVisualization/index.js
@@ -1,7 +1,10 @@
 import { i18n } from 'vizhub-i18n';
 import { zipFiles } from './zipFiles';
-import vizhubLibraries from 'vizhub-libraries';
-import { getFileIndex, dependencies } from 'vizhub-presenters';
+import {
+  getFileIndex,
+  getComputedIndexHtml,
+  getLibraries,
+} from 'vizhub-presenters';
 
 export class ExportVisualization {
   constructor({ visualizationGateway }) {
@@ -17,28 +20,27 @@ export class ExportVisualization {
       id: requestModel.id,
     });
 
-    const packageJSONDependencies = dependencies(visualization.content.files);
-    const packageJSONDependencyNames = Object.keys(packageJSONDependencies);
+    const libraries = getLibraries(visualization.content.files);
 
-    const globals = packageJSONDependencyNames.reduce(
-      (globals, packageName) => {
-        if (vizhubLibraries[packageName]) {
-          delete globals[packageName];
-        }
-
-        return globals;
-      },
-      { ...vizhubLibraries }
-    );
-
-    const globalsJSON = JSON.stringify(globals);
-    const externalJSON = JSON.stringify(Object.keys(globals));
+    const globalsJSON = JSON.stringify(libraries);
+    const externalJSON = JSON.stringify(Object.keys(libraries));
 
     const files = visualization.content.files.concat([
       {
+        name: 'package.json',
+        text: `{
+  "scripts": {
+    "build": "rollup -c"
+  },
+  "devDependencies": {
+    "rollup": "latest",
+    "@rollup/plugin-buble": "latest"
+  }
+}`,
+      },
+      {
         name: 'rollup.config.js',
         text: `const buble = require('@rollup/plugin-buble');
-  const { nodeResolve } = require('@rollup/plugin-node-resolve');
   
   export default {
   input: 'index.js',
@@ -49,33 +51,13 @@ export class ExportVisualization {
     sourcemap: true,
     globals: ${globalsJSON}
   },
-  plugins: [nodeResolve(), buble({exclude: ['node_modules/**/*']})]
+  plugins: [buble()]
 };`,
       },
     ]);
 
-    const packageJSONFile = {
-      name: 'package.json',
-      text: `{
-"scripts": {
-  "build": "rollup -c"
-},
-"dependencies": ${JSON.stringify(packageJSONDependencies)},
-"devDependencies": {
-  "rollup": "latest",
-  "@rollup/plugin-buble": "latest",
-  "@rollup/plugin-node-resolve": "latest"
-}
-}`,
-    };
-
-    const packageJSONFileIndex = getFileIndex(files, 'package.json');
-
-    if (packageJSONFileIndex === -1) {
-      files.append(packageJSONFile);
-    } else {
-      files[packageJSONFileIndex] = packageJSONFile;
-    }
+    const indexHtmlFileIndex = getFileIndex(files, 'index.html');
+    files[indexHtmlFileIndex].text = getComputedIndexHtml(files);
 
     const zipFileBuffer = zipFiles(files);
     const zipFileName = visualization.info.title + '.zip';
